chore(backend): Merge test utils to util/testUtils.ts module
