+++ packages/backend/src/tokens/__tests__/verify.test.ts
@@ -2,7 +2,7 @@ import type QUnit from 'qunit';
 import sinon from 'sinon';
 
 import runtime from '../../runtime';
-import { jsonOk } from '../../util/mockFetch';
+import { jsonOk } from '../../util/testUtils';
 import { mockJwks, mockJwt, mockJwtPayload } from '../fixtures';
 import { verifyToken } from '../verify';
